refactor: [MON-260] 시리즈 생성, form 로직을 SeriesForm으로 통합
